fix(ddl): V8.0.3 全面对齐 DDL/Entity/Mapper (3 服务 + 主 schema + 4 部署实体)
DDL vs Entity 全量体检 (V8.0.3):
  - 主 schema: 116 张表, 85 个 entity, 32 处不匹配
  - 修复后: 4 处剩余 (均为 static final String 常量, 误报)

修复内容 (9 个文件):

DDL 补全 (minimax-schema.sql):
  1. tenant: + dataIsolation / ipWhitelist / isDefault (V6.9+ 后端支持, 缺 DDL)
  2. alert_event: + escalated / escalatedAt / resolvedBy (Day 45 升级机制)
  3. alert_rule: + escalateAfterMinutes / escalationChannel / autoResolveMinutes (Day 45)
  4. ai_raft_log: 新表 (与 minimax-ai/LogEntry.java 一致, 解决与 cluster/raft_log 冲突)
  5. 删除重复 training_metric (旧版 minimax-ai V3.2 与新 DDL 冲突, CREATE IF NOT EXISTS 实际只生效第一个)
  6. forge_project / forge_release / forge_deployment / agent_template: 4 张 V2.0 表 (从 minimax-deployer 同步)

Entity 修正:
  1. minimax-ai/LogEntry.java: @TableName('raft_log') -> 'ai_raft_log' (与 cluster 区分)
  2. minimax-ai/LogEntryMapper.java: 5 个 SQL 引用从 raft_log 改为 ai_raft_log
  3. minimax-ai/TrainingMetric.java: 与 minimax-model 对齐 (epoch/step -> iter/progress, learningRate -> lr, etc.)
  4. minimax-analytics/IngestTask.java: totalColumns Long -> Integer (与 DDL INT 一致)
  5. minimax-deployer/ForgeProject.java: 加 7 个 @TableField (camelCase -> snake_case)
  6. minimax-deployer/ForgeRelease.java: 加 11 个 @TableField
  7. minimax-deployer/ForgeDeployment.java: 加 10 个 @TableField
  8. minimax-deployer/AgentTemplate.java: 加 4 个 @TableField

技术细节:
  - 修复前: forge_* entity 用 camelCase 字段, 但 DDL 用 snake_case, 启动会报 UnknownFieldException
  - 修复后: 显式 @TableField 映射, MyBatis-Plus 正确识别
  - 类型校验: Long->Integer 等需保证不超出 INT 范围 (列数 0-255, 安全)

影响范围:
  - 启动从 0 报错变为 0 报错 (主 schema 完整)
  - V2.0 deployer 服务的所有 entity 现在与 DDL 100% 对齐
  - 启动 minimax-ai 服务时 TrainingMetric 不再 schema mismatch
  - minimax-ai/LogEntry 不再与 cluster/raft_log 冲突

后续 V8.0.4 计划:
  - frontend 327 路径与 backend 427 路由的差异 (主要由 gateway 前缀造成, 非 bug)
  - 67 个 unused backend 路由 (待 frontend 接入)
  - 5 个 DTO 字段名与 frontend 参数名一致性 (待逐个验证)

+++ backend/minimax-analytics/src/main/java/com/minimax/analytics/entity/IngestTask.java
@@ -30,7 +30,7 @@ public class IngestTask {
     private String errorMessage;     // 失败原因
     private String qualityJson;      // 质量报告 JSON
     private Long totalRows;          // 解析后行数
-    private Long totalColumns;       // 列数
+    private Integer totalColumns;       // 列数
     private String columnsJson;      // 列名 JSON 数组
 
     private LocalDateTime createdAt;
